Should've committed more often, oh well, this contains two changes - is_number -> AST_num and converted all tabs to spaces, diff with git's --ignore-all-space. I wrote untabify because Visual Studio wouldn't do it correctly.

+++ ++c/interp.h
@@ -3,4 +3,4 @@
 
 // a simple interpreter that will take an AST and either fail to execute or return a final result
 
-bool interp_return_value(ASTNode* root, int* out_result);
+bool interp_return_value(ASTNode* root, int64_t* out_result);
